Reorganize signal tests
