+++ packages/liveblocks-core/src/lib/__tests__/signals/Signal.test.ts
@@ -1,7 +1,7 @@
 import fc from "fast-check";
 
-import { shallow } from "../../lib/shallow";
-import { batch, Signal } from "../Signal";
+import { shallow } from "../../../lib/shallow";
+import { batch, Signal } from "../../signals";
 
 test("empty", () => {
   expect(new Signal({}).get()).toStrictEqual({});
